Rename actor package to operation
